requirements.txt
numpy==1.16.3
pandas==0.23.4
Flask==1.0.2
gunicorn==19.9.0
scipy==1.2.1
google-cloud-bigquery==1.11.2
requests==2.21.0